I've made 2 tests with the debugger.. Just below is the actual intentSettingsActivity: (intentSettingsActivity = new Intent();) intentSettingsActivity = {Intent@830021487784} "Intent { act=com.example.test1pn.myAction1 }"  mAction = "com.example.test1pn.myAction1"  mCategories = null  mComponent = null  mData = null  mExtras = null  mType = null  mPackage = null  mSelector = null  mSourceBounds = null  mFlags = 0 pendingIntent = null
//Vl. below is the same intentSettingsActivity after creation with: intentSettingsActivity = new Intent(maContext, SettingsActivity.class); //Vl.explicit intent
intentSettingsActivity = {Intent@830033960944} "Intent { act=com.example.test1pn.myAction1 cmp=xxxx }"
 mAction = "com.example.test1pn.myAction1"
 mCategories = null
 mComponent = {ComponentName@830033961000} "ComponentInfo{com.example.test1pn/com.example.test1pn.SettingsActivity}"
  mClass = "com.example.test1pn.SettingsActivity"
   value = {char[36]@830032906744}
   hashCode = 504443124
   offset = 0
   count = 36
  mPackage = "com.example.test1pn"
 mData = null
 mExtras = null
 mType = null
 mPackage = null
 mSelector = null
 mSourceBounds = null
 mFlags = 0
pendingIntent = {PendingIntent@830032919592} "PendingIntent{41d13c28: android.os.BinderProxy@41d13bb8}"

//Vl. as we can see the main difference is in the mComponent field..

Finally the explanation (https://developer.android.com/reference/android/content/Intent.html):
component -- Specifies an explicit name of a component class to use for the intent. Normally this is determined by looking at
the other information in the intent (the action, data/type, and categories) and matching that with a component that can handle it.
If this attribute is set then none of the evaluation is performed, and this component is used exactly as is. By specifying this attribute,
all of the other Intent attributes become optional.
//Vl. that's why previously "intentSettingsActivity = new Intent(maContext, SettingsActivity.class" didn't match our intent filter.. just because
when mComponent (secondary attribute) is not null, "all of the other Intent attributes become optional" (including main attributes:
action & data).