Now when we press 12.Intents1 button, a chooser window appear.. as below Title: Complete action using 1. Gallery 2. Gallery(Only View) 	Use by default for this action (check box) I always choose Gallery(Only View).. Then I select a picture in Gallery(Only View).., and we get e.g: Vladi5, FileURI: content://media/external/images/media/5274
Interesting remark:
When we had:
Intent pickIntent = new Intent(Intent.ACTION_PICK);
the chooser was:
1. SIM Toolkit
2. WhatsApp
Now with:
Intent pickIntent = new Intent(Intent.ACTION_PICK, android.provider.MediaStore.Images.Media.EXTERNAL_CONTENT_URI);
the chooser is:
1. Gallery
2. Gallery(Only View)
I think this is because android.intent.action.PICK is a MIME Typed Action..
MIME Typed Actions:
      android.intent.action.ATTACH_DATA:
        411d4138 com.android.gallery3d/.app.Wallpaper filter 411d4260
      com.android.camera.action.CROP:
        411d4818 com.android.gallery3d/.app.CropImage filter 411d4aa8
      android.intent.action.GET_CONTENT:
        411d2600 com.android.gallery3d/.app.Gallery filter 411d29d0
        411d2600 com.android.gallery3d/.app.Gallery filter 411d2c98
      android.intent.action.PICK:
        411d2600 com.android.gallery3d/.app.Gallery filter 411d36b0
        411d2600 com.android.gallery3d/.app.Gallery filter 411d38c0
      android.intent.action.EDIT:
        411d4e58 com.android.gallery3d/.photoeditor.PhotoEditor filter 411d4fe0
      com.android.camera.action.REVIEW:
        411d2600 com.android.gallery3d/.app.Gallery filter 411d30a8
        411d2600 com.android.gallery3d/.app.Gallery filter 411d34e8

so,.. I think 1st instance / element of the filter (<action>, <data>, or <category>) is the MIME data type and 2nd is ACTION..
It would be interesting to see what behavior we get with an intent with data only (without action)..

https://developer.android.com/guide/components/intents-filters.html :
https://developer.android.com/guide/components/intents-filters.html#Resolution
If the filter does not list any actions, there is nothing for an intent to match, so all intents fail the test. However, if an Intent
does not specify an action, it passes the test as long as the filter contains at least one action.